Replaces swap chain lazily when the swap interval is changed.
It does not appear to be possible to pass a presentation interval to Present() or change it after creating the swap chain, short of calling Reset(). I decided to try fixing it by replacing the swap chain under the same conditions as a window resize.

I tested it with a modified version of the simple vertex shader sample, passing various values to eglSwapInterval.

My reading of the EGL spec might be wrong but I believe the interval state should be associated with the surface bound to the current context, rather than the display.



Review URL: http://codereview.appspot.com/2036044

git-svn-id: http://angleproject.googlecode.com/svn/trunk@408 736b8ea6-26fd-11df-bfd4-992fa37f6226