Changed raw pointers to RefCountObject to BindingPointers.
I suspect there is a bug whereby an owning reference to a RefCountObject is temporarily assigned to a BindingPointer, causing it to be destroyed and making the owning raw pointer go dangling.

Making this change for mColorbufferProxy seemed to fix a crash in Chrome with canvas 2D.

This was previously landed in r417 and r418.

Review URL: http://codereview.appspot.com/2108047

git-svn-id: http://angleproject.googlecode.com/svn/trunk@423 736b8ea6-26fd-11df-bfd4-992fa37f6226