Adding code to try and reset the device if eglCreateWindowSurface and eglCreateSurface are called while the device is reported as lost.
Currently only eglCreateContext attempts to reset the device. This is a short term fix until the whole device lost issue can be revisited.  One of the issues here is that if any of these egl entry points are called while the device is lost (e.g. while the screen saver is on in XP) ANGLE will spin until the device comes back.
Review URL: http://codereview.appspot.com/4961070

git-svn-id: http://angleproject.googlecode.com/svn/trunk@746 736b8ea6-26fd-11df-bfd4-992fa37f6226