Ensure texture storage exists before retrieving serial
Trac #19330
Signed-off-by: Nicolas Capens

In some cases we were trying to retrieve a serial from
a rendertarget texture which didn't have the storage allocated yet.
This was resulting in a 0 serial which was indistinguishable from
the case where context was just marked dirty and we assumed
the target was up-to-date. This resulted in
failing to apply the correct rendertarget and as a result
some d3d clear calls were failing (picked up by debug runtime).

git-svn-id: http://angleproject.googlecode.com/svn/trunk@929 736b8ea6-26fd-11df-bfd4-992fa37f6226