Make Image::lock and unlock private and call them from Image::loadData.
This is because I want loadData to be able to decide whether to use LockRect/UnlockRect or UpdateSurface as the method to load the data, depending on whether the Images are backed by textures in the default pool or not.

I think it actually cleans up the code a bit.

Also, fixed a bug in subImageCompressed. The associated WebGL conformance test now passes.
Review URL: https://codereview.appspot.com/5948043

git-svn-id: http://angleproject.googlecode.com/svn/trunk@1013 736b8ea6-26fd-11df-bfd4-992fa37f6226